# [copy] improve copy perf