# After

```
root@C.18059546:~/misc-cuda/src$ make && ./bin/copy
make: Nothing to be done for 'all'.
Thread 0 took 6 cycles per element
Thread 1 took 6 cycles per element
Thread 2 took 6 cycles per element
Thread 3 took 6 cycles per element
Thread 4 took 6 cycles per element
Thread 5 took 6 cycles per element
Thread 6 took 6 cycles per element
Thread 7 took 6 cycles per element
Thread 8 took 6 cycles per element
Thread 9 took 6 cycles per element
Thread 10 took 6 cycles per element
Thread 11 took 6 cycles per element
Thread 12 took 6 cycles per element
Thread 13 took 6 cycles per element
Thread 14 took 6 cycles per element
Thread 15 took 6 cycles per element
Thread 16 took 6 cycles per element
Thread 17 took 6 cycles per element
Thread 18 took 6 cycles per element
Thread 19 took 6 cycles per element
Thread 20 took 6 cycles per element
Thread 21 took 6 cycles per element
Thread 22 took 6 cycles per element
Thread 23 took 6 cycles per element
Thread 24 took 6 cycles per element
Thread 25 took 6 cycles per element
Thread 26 took 6 cycles per element
Thread 27 took 6 cycles per element
Thread 28 took 6 cycles per element
Thread 29 took 6 cycles per element
Thread 30 took 6 cycles per element
Thread 31 took 6 cycles per element
Time taken: 7.188480 ms
```

# Before

```
make && ./bin/copy
nvcc -O2 -arch=sm_80 -o bin/copy copy.cu
Thread 0 took 555 cycles per element
Thread 1 took 555 cycles per element
Thread 2 took 555 cycles per element
Thread 3 took 555 cycles per element
Thread 4 took 555 cycles per element
Thread 5 took 555 cycles per element
Thread 6 took 555 cycles per element
Thread 7 took 555 cycles per element
Thread 8 took 555 cycles per element
Thread 9 took 555 cycles per element
Thread 10 took 555 cycles per element
Thread 11 took 555 cycles per element
Thread 12 took 555 cycles per element
Thread 13 took 555 cycles per element
Thread 14 took 555 cycles per element
Thread 15 took 555 cycles per element
Thread 16 took 555 cycles per element
Thread 17 took 555 cycles per element
Thread 18 took 555 cycles per element
Thread 19 took 555 cycles per element
Thread 20 took 555 cycles per element
Thread 21 took 555 cycles per element
Thread 22 took 555 cycles per element
Thread 23 took 555 cycles per element
Thread 24 took 555 cycles per element
Thread 25 took 555 cycles per element
Thread 26 took 555 cycles per element
Thread 27 took 555 cycles per element
Thread 28 took 555 cycles per element
Thread 29 took 555 cycles per element
Thread 30 took 555 cycles per element
Thread 31 took 555 cycles per element
Time taken: 11.256448 ms
```